# backend/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RefuLearn API",
    "description": "A comprehensive API for the RefuLearn platform supporting refugee education and skill development",
    "version": "1.0.0",
    "contact": {
      "name": "RefuLearn Team",
      "email": "support@refulearn.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.refulearn.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User ID"
          },
          "firstName": {
            "type": "string",
            "description": "User's first name"
          },
          "lastName": {
            "type": "string",
            "description": "User's last name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "role": {
            "type": "string",
            "enum": ["refugee", "instructor", "admin", "employer"],
            "description": "User's role in the platform"
          },
          "profilePic": {
            "type": "string",
            "description": "URL to user's profile picture"
          },
          "country": {
            "type": "string",
            "description": "User's country"
          },
          "city": {
            "type": "string",
            "description": "User's city"
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Languages the user speaks"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "User's skills"
          },
          "interests": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "User's interests"
          },
          "isEmailVerified": {
            "type": "boolean",
            "description": "Whether user's email is verified"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether user account is active"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation date"
          }
        }
      },
      "Course": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Course ID"
          },
          "title": {
            "type": "string",
            "description": "Course title"
          },
          "description": {
            "type": "string",
            "description": "Course description"
          },
          "overview": {
            "type": "string",
            "description": "Detailed course overview"
          },
          "level": {
            "type": "string",
            "enum": ["Beginner", "Intermediate", "Advanced"],
            "description": "Course difficulty level"
          },
          "duration": {
            "type": "string",
            "description": "Course duration (e.g., '8 weeks')"
          },
          "category": {
            "type": "string",
            "description": "Course category"
          },
          "instructor_id": {
            "type": "string",
            "description": "Instructor ID"
          },
          "modules": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Course modules"
          },
          "learningObjectives": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Learning objectives"
          },
          "image": {
            "type": "string",
            "description": "Course image URL"
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether course is active"
          },
          "isPublished": {
            "type": "boolean",
            "description": "Whether course is published"
          },
          "course_profile_picture": {
            "type": "string",
            "format": "binary",
            "description": "Course profile picture"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Course creation date"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Course last update date"
          }
        },
        "required": ["_id", "title", "category", "instructor_id", "is_active", "isPublished", "course_profile_picture"]
      },
...
      "Certificate": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "description": "Certificate ID" },
          "student": { "$ref": "#/components/schemas/User" },
          "course": { "$ref": "#/components/schemas/Course" },
          "instructor": { "$ref": "#/components/schemas/User" },
          "certificateNumber": { "type": "string", "description": "Unique certificate number" },
          "title": { "type": "string", "description": "Certificate title" },
          "description": { "type": "string", "description": "Certificate description" },
          "issueDate": { "type": "string", "format": "date-time", "description": "Issue date" },
          "expiryDate": { "type": "string", "format": "date-time", "description": "Expiry date" },
          "status": { "type": "string", "enum": ["active", "expired", "revoked"], "description": "Certificate status" },
          "grade": { "type": "string", "description": "Grade achieved" },
          "score": { "type": "number", "description": "Score achieved" },
          "maxScore": { "type": "number", "description": "Maximum possible score" },
          "percentage": { "type": "number", "description": "Percentage achieved" },
          "completionDate": { "type": "string", "format": "date-time", "description": "Course completion date" },
          "certificateUrl": { "type": "string", "description": "URL to certificate file" },
          "verificationCode": { "type": "string", "description": "Unique verification code" },
          "isVerified": { "type": "boolean", "description": "Whether certificate is verified" },
          "verifiedAt": { "type": "string", "format": "date-time", "description": "Verification date" },
          "verifiedBy": { "$ref": "#/components/schemas/User" },
          "achievements": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": { "type": "string" },
                "description": { "type": "string" },
                "date": { "type": "string", "format": "date-time" }
              }
            }
          },
          "skills": {
            "type": "array",
            "items": { "type": "string" }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "courseDuration": { "type": "string" },
              "totalModules": { "type": "number" },
              "completedModules": { "type": "number" },
              "totalAssessments": { "type": "number" },
              "passedAssessments": { "type": "number" },
              "timeSpent": { "type": "number" }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
    },
  "paths": {
...
    "/api/courses/certificates": {
      "get": {
        "tags": ["courses"],
        "summary": "List all certificates",
        "description": "List all certificates (admin/instructor only, filter by student, course, or status)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "student", "in": "query", "schema": { "type": "string" }, "description": "Filter by student ID" },
          { "name": "course", "in": "query", "schema": { "type": "string" }, "description": "Filter by course ID" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["active", "expired", "revoked"] }, "description": "Filter by status" }
        ],
        "responses": {
          "200": {
            "description": "Certificates retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "certificates": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Certificate" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": { "description": "Forbidden" }
        }
      },
      "post": {
        "tags": ["courses"],
        "summary": "Create a certificate",
        "description": "Create a new certificate (instructor/admin only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["student", "course", "title", "description", "completionDate"],
                "properties": {
                  "student": { "type": "string", "description": "Student ID" },
                  "course": { "type": "string", "description": "Course ID" },
                  "title": { "type": "string", "description": "Certificate title" },
                  "description": { "type": "string", "description": "Certificate description" },
                  "completionDate": { "type": "string", "format": "date-time", "description": "Certificate completion date" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Certificate created successfully" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/api/courses/certificates/{certificateId}": {
      "get": {
        "tags": ["courses"],
        "summary": "Get certificate by ID",
        "description": "Get a certificate by its ID (only user, admin, or instructor)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "certificateId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Certificate retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "certificate": { "$ref": "#/components/schemas/Certificate" }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": { "description": "Forbidden" },
          "404": { "description": "Certificate not found" }
        }
      },
      "patch": {
        "tags": ["courses"],
        "summary": "Update certificate",
        "description": "Update a certificate (only admin or instructor)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "certificateId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["student", "course", "title", "description", "completionDate", "status"],
                "properties": {
                  "student": { "type": "string", "description": "Student ID" },
                  "course": { "type": "string", "description": "Course ID" },
                  "title": { "type": "string", "description": "Certificate title" },
                  "description": { "type": "string", "description": "Certificate description" },
                  "completionDate": { "type": "string", "format": "date-time", "description": "Certificate completion date" },
                  "status": { "type": "string", "description": "Certificate status" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Certificate updated successfully" },
          "404": { "description": "Certificate not found" }
        }
      },
      "delete": {
        "tags": ["courses"],
        "summary": "Delete certificate",
        "description": "Delete a certificate (only admin or instructor)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "certificateId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Certificate deleted successfully" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Certificate not found" }
        }
      }
    },
...
    "/api/admin/analytics": {
      "get": {
        "tags": ["admin"],
        "summary": "Get admin analytics",
        "description": "Retrieve comprehensive analytics data for admin dashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Analytics data retrieved successfully" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/instructor/analytics": {
      "get": {
        "tags": ["instructor"],
        "summary": "Get instructor analytics",
        "description": "Retrieve comprehensive analytics for the instructor",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Analytics retrieved successfully" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  }
}